requirements.txt
altgraph==0.17.4
Nuitka==2.3.11
ordered-set==4.1.0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
zstandard==0.23.0
